+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Support --locale command line option
+
 # 0.1.0
 
 * Initial release.